// proto/projection/ProjectionContext.ts
// Original file: proto/project.proto


export interface ProjectionContext {
  'eventType'?: (string);
  'eventId'?: (string);
  'stream'?: (string);
  'eventJson'?: (string);
}

export interface ProjectionContext__Output {
  'eventType': (string);
  'eventId': (string);
  'stream': (string);
  'eventJson': (string);
}